backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
pydantic-settings==2.3.1
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
asyncio==3.4.3
python-multipart==0.0.9
opencv-python==4.10.0.84
ultralytics==8.2.32
openai-whisper==20231117
librosa==0.10.2
spacy==3.7.5
python-dotenv==1.0.1
aiofiles==23.2.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
